split traits and factory into files
